Menu: extract the System tab into build_system.cpp
Verbatim move of the System sections (headset, audio, timers & alarm,
demo/diagnostics, HUD/menu presets, GPIO visualizer + buttons, Pi
settings, updater, power, display, connectivity, profiles/software and
the System root). push_notif and make_builtin_theme_leaves are now
constructed once at the top of build_menu and shared through the
context; the Phone (KDE Connect) item crosses from the Communications
block to System > Connectivity via ctx.phone_item. run_i2c_scan moves
along (System-only).
