+++ CMakeLists.txt
@@ -204,6 +204,7 @@ set(SOURCES
     src/hud/background_library.cpp
     src/menu/build_menu.cpp
     src/menu/build_quick.cpp
+    src/menu/build_system.cpp
     src/menu/menu_system.cpp
     src/menu/file_picker.cpp
     src/menu/face_editor.cpp
